performance-test/requirements.txt
locustio==0.9.0
invokust==0.5.2
pytest==5.3.1